src/sub/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
openai==0.27.8
packaging==23.1
pluggy==1.2.0
pytest==7.4.0
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.4.1
tqdm==4.65.0
urllib3==2.0.4
wikipedia==1.4.0
yarl==1.9.2
google-cloud-storage==1.23.0
